# api/openapi.yaml
openapi: 3.1.0
info:
  title: Forever Sixty API
  version: 0.1.0
servers:
  - url: https://api.foreversixty.gg
components:
  schemas:
    Envelope:
      type: object
      required: [ok, data, error, request_id]
      properties:
        ok: { type: boolean }
        data:
          description: any JSON or null
        error:
          oneOf:
            - type: "null"
            - type: object
              properties:
                code: { type: string }
                message: { type: string }
                fields: { type: object, additionalProperties: { type: string } }
        request_id: { type: string }
  responses:
    NotFound:
      description: "Unmatched routes return the envelope with code not_found"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
paths:
  /healthz:
    get:
      summary: Liveness
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /version:
    get:
      summary: Build version
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /v1/subscribe:
    post:
      summary: Start a double opt-in subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '202':
          description: "Accepted regardless of whether the address is already known"
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '400':
          description: "Malformed or missing JSON body, or invalid email address. When email is invalid, error.fields.email is set."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '429':
          description: "Rate limited: 10 requests per minute per IP on this route, 120 per minute per IP overall"
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '500':
          description: "Unexpected store error; envelope with code internal. Mail is sent asynchronously after the response, so this is never a mail-provider failure."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /v1/subscribe/confirm:
    get:
      summary: Confirm a subscription
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '302':
          description: "Redirect to configured public site base URL + /subscribed (success) or /subscribe-invalid (invalid token)"
  /v1/subscribe/unsubscribe:
    get:
      summary: Unsubscribe
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '302':
          description: "Redirect to configured public site base URL + /unsubscribed (success) or /subscribe-invalid (invalid token)"